+++ src/test/java/executor/service/steps/ClickCssTest.java
@@ -1,4 +1,4 @@
-package executor.service;
+package executor.service.steps;
 
 import executor.service.model.StepDTO;
 import org.junit.jupiter.api.BeforeEach;
